+++ src/angular/mod.rs
@@ -1,4 +1,9 @@
+#[cfg(all(unix, feature = "background"))]
+mod background;
 mod builder;
 mod workspace;
 
 pub(crate) use builder::build;
+
+#[cfg(all(unix, feature = "background"))]
+pub(crate) use background::stop as stop_background_process;