+++ src/js/experimental-builder/builder.mjs
@@ -5,47 +5,49 @@ import {createBuilder} from '@angular-devkit/architect';
 import fs from 'node:fs';
 import {posix} from 'node:path';
 
-export default createBuilder(({optimization, ...input}, context) => {
-	return buildApplicationInternal(
-		{
-			watch: false,
-			progress: false,
+export default createBuilder(
+	({optimization, watch = false, ...input}, context) => {
+		return buildApplicationInternal(
+			{
+				watch,
+				progress: false,
 
-			index: false,
-			entryPoints: new Set(
-				fs
-					.readdirSync(context.workspaceRoot)
-					.filter(file => {
-						let resolvedFile = posix.join(context.workspaceRoot, file);
+				index: false,
+				entryPoints: new Set(
+					fs
+						.readdirSync(context.workspaceRoot)
+						.filter(file => {
+							let resolvedFile = posix.join(context.workspaceRoot, file);
 
-						return (
-							fs.statSync(resolvedFile).isDirectory() &&
-							fs.existsSync(`${resolvedFile}/${file}.ts`)
-						);
-					})
-					.map(file => `${file}/${file}.ts`),
-			),
-			aot: true,
+							return (
+								fs.statSync(resolvedFile).isDirectory() &&
+								fs.existsSync(`${resolvedFile}/${file}.ts`)
+							);
+						})
+						.map(file => `${file}/${file}.ts`),
+				),
+				aot: true,
 
-			tsConfig: 'tsconfig.json',
-			...input,
+				tsConfig: 'tsconfig.json',
+				...input,
 
-			...(optimization
-				? {
-						optimization: {
-							styles: {
-								inlineCritical: false,
-								minify: true,
+				...(optimization
+					? {
+							optimization: {
+								styles: {
+									inlineCritical: false,
+									minify: true,
+								},
+								scripts: true,
 							},
-							scripts: true,
-						},
-						outputHashing: 'all',
-				  }
-				: {
-						optimization: false,
-						outputHashing: 'none',
-				  }),
-		},
-		context,
-	);
-});
+							outputHashing: 'all',
+					  }
+					: {
+							optimization: false,
+							outputHashing: 'none',
+					  }),
+			},
+			context,
+		);
+	},
+);